Use cdivision to speedup modulo operation
If set to False, Cython will adjust the remainder and quotient operators
C types to match those of Python ints (which differ when the operands
have opposite signs) and raise a ZeroDivisionError when the right
operand is 0. This has up to a 35% speed penalty. If set to True, no
checks are performed. See CEP 516. Default is False.