Implement bit counting in BitField
Counting number of set bits in a byte (BitField) allows to
count number of set bits in BitVector and, consequently,
in BloomFilter, which is important for an estimation
of number of added elements into the filter.

We use, probabily, the fastest algorithm to count bits
in a byte - predefined lookup table for each possible
values out of 256.