Fix cardinality estimation in Bloom filters
When filter is full the error of the cardinality estimation
is quite big and usually significantly underfits the real values.
However, our estimation was not correct (actually, it may correct
when only 1 hash function is used, which is unlikely the case).

In case the filter is full, we assume that no coolisions happend,
menaing that the ratio of the filter's length to the number of
used hash functions provides the estimation.